requirements.txt
# Core
langgraph
langchain
langchain-openai
langchain-core
langchain-community
langchain-groq

# Pydantic
pydantic
pydantic[email]
pydantic-settings

# PDF Processing
PyPDF2  

# Document Generation
python-docx
reportlab
markdown

# NLP & Text Processing
spacy
nltk

# Data Processing
pandas
numpy

# Utilities
python-dotenv

# UI    
streamlit   
gradio
plotly 
streamlit-aggrid
streamlit-extras

# # Web Search & Scraping
duckduckgo-search
requests
beautifulsoup4

# # GitHub API
PyGithub

# # Caching for API calls
requests-cache

# # Additional utilities
python-dateutil
fuzzywuzzy
python-Levenshtein

ruff
